Change silence state duration models
In earlier times, all model names where lower case. Because of this, silence
states (starting with an underscore) are always sorted on top. This allowed
for tricks like "skip-first-n-states" when doing duration modeling.

Nowadays, we use uppercase model names for funny characters. To replace the
earlier behaviour, we now afterwards remove the duration models for every
state (which has a centerphone) that starts with an _. It is done with perl
grep and sed magic.

I am not proud of this, but hey, it works. Anyone that could rewrite our
scripts into a sensible language, please go ahead.

I'll take a shower now.